cleaned up the project
