+++ nctts/homepage/templates/authenticate/gov-login.html
@@ -40,7 +40,7 @@
   <div class="submit-row">
     <input type="submit" value="{% translate 'Login' %}">
   </div>
-  <p><a href="{% url 'password_reset' %}">Lost password?</a></p>
+  <p><a href="{% url 'password_reset' %}">Lost password?xxxx</a></p>
 </form>
