+++ nctts/homepage/templates/homepage/home.html
@@ -27,7 +27,7 @@
                     {% endblock %}
                 </ul>
                 {% block login %}
-                <a class="btn btn-primary ms-md-2" role="button" href="{% url 'gov-login' %}">Login</a>
+                <a class="btn btn-primary ms-md-2" role="button" href="{% url 'homepage:login_user' %}">Login</a>
                 {% endblock %}
             </div>
         </div>
