+++ nctts/homepage/urls.py
@@ -4,5 +4,8 @@ from . import views
 app_name = 'homepage'
 
 urlpatterns = [
-    path('', views.HomeView.as_view(), name='home')
+    path('', views.HomeView.as_view(), name='home'),
+    path('login_user', views.login_user, name='login_user'),
+    path('login_success', views.login_success, name='login_success'),
+    path('login_failed', views.login_failed, name='login_failed')
 ]
