+++ nctts/nctts/urls.py
@@ -18,12 +18,9 @@ from django.contrib import admin
 from django.urls import path, include
 
 urlpatterns = [
-    path("admin/", admin.site.urls, name='admin'),
-    path("sendEmail/", include("sendEmail.urls")),
-    path("report/", include("report.urls"), name='report'),
     path("", include('homepage.urls')),
-    path("gov/", include("gov.urls"), name='gov-login'),
-    path("gov/", include("django.contrib.auth.urls")),
-    path('operatorview/', include("operatorview.urls")),
-    path('accounts/', include('django.contrib.auth.urls'))
+    path("", include("django.contrib.auth.urls")),
+    path("admin/", admin.site.urls),
+    path("report/", include("report.urls")),
+    path('operatorview/', include("operatorview.urls"))
 ]
